refactor: reposition useTheme hook to inside hooks folder
