+++ src/hooks/use-theme.tsx
@@ -1,0 +1,11 @@
+import { ThemeProviderContext } from "@/components/theme-provider";
+import { useContext } from "react";
+
+export const useTheme = () => {
+  const context = useContext(ThemeProviderContext);
+
+  if (context === undefined)
+    throw new Error("useTheme must be used within a ThemeProvider");
+
+  return context;
+};
